Extract date/time template functions into an datetime Jinja2 extension
