+++ homeassistant/helpers/template/__init__.py
@@ -7,7 +7,7 @@ import asyncio
 import collections.abc
 from collections.abc import Callable, Generator, Iterable
 from copy import deepcopy
-from datetime import date, datetime, time, timedelta
+from datetime import datetime, timedelta
 from functools import cache, lru_cache, partial, wraps
 import json
 import logging
@@ -63,7 +63,7 @@ from homeassistant.helpers import (
 from homeassistant.helpers.singleton import singleton
 from homeassistant.helpers.translation import async_translate_state
 from homeassistant.helpers.typing import TemplateVarsType
-from homeassistant.util import convert, dt as dt_util, location as location_util
+from homeassistant.util import convert, location as location_util
 from homeassistant.util.async_ import run_callback_threadsafe
 from homeassistant.util.hass_dict import HassKey
 from homeassistant.util.json import JSON_DECODE_EXCEPTIONS, json_loads
@@ -1417,22 +1417,6 @@ def has_value(hass: HomeAssistant, entity_id: str) -> bool:
     )
 
 
-def now(hass: HomeAssistant) -> datetime:
-    """Record fetching now."""
-    if (render_info := render_info_cv.get()) is not None:
-        render_info.has_time = True
-
-    return dt_util.now()
-
-
-def utcnow(hass: HomeAssistant) -> datetime:
-    """Record fetching utcnow."""
-    if (render_info := render_info_cv.get()) is not None:
-        render_info.has_time = True
-
-    return dt_util.utcnow()
-
-
 def forgiving_round(value, precision=0, method="common", default=_SENTINEL):
     """Filter to round a value."""
     try:
@@ -1510,85 +1494,6 @@ def version(value):
     return AwesomeVersion(value)
 
 
-def timestamp_custom(value, date_format=DATE_STR_FORMAT, local=True, default=_SENTINEL):
-    """Filter to convert given timestamp to format."""
-    try:
-        result = dt_util.utc_from_timestamp(value)
-
-        if local:
-            result = dt_util.as_local(result)
-
-        return result.strftime(date_format)
-    except (ValueError, TypeError):
-        # If timestamp can't be converted
-        if default is _SENTINEL:
-            raise_no_default("timestamp_custom", value)
-        return default
-
-
-def timestamp_local(value, default=_SENTINEL):
-    """Filter to convert given timestamp to local date/time."""
-    try:
-        return dt_util.as_local(dt_util.utc_from_timestamp(value)).isoformat()
-    except (ValueError, TypeError):
-        # If timestamp can't be converted
-        if default is _SENTINEL:
-            raise_no_default("timestamp_local", value)
-        return default
-
-
-def timestamp_utc(value, default=_SENTINEL):
-    """Filter to convert given timestamp to UTC date/time."""
-    try:
-        return dt_util.utc_from_timestamp(value).isoformat()
-    except (ValueError, TypeError):
-        # If timestamp can't be converted
-        if default is _SENTINEL:
-            raise_no_default("timestamp_utc", value)
-        return default
-
-
-def forgiving_as_timestamp(value, default=_SENTINEL):
-    """Filter and function which tries to convert value to timestamp."""
-    try:
-        return dt_util.as_timestamp(value)
-    except (ValueError, TypeError):
-        if default is _SENTINEL:
-            raise_no_default("as_timestamp", value)
-        return default
-
-
-def as_datetime(value: Any, default: Any = _SENTINEL) -> Any:
-    """Filter and to convert a time string or UNIX timestamp to datetime object."""
-    # Return datetime.datetime object without changes
-    if type(value) is datetime:
-        return value
-    # Add midnight to datetime.date object
-    if type(value) is date:
-        return datetime.combine(value, time(0, 0, 0))
-    try:
-        # Check for a valid UNIX timestamp string, int or float
-        timestamp = float(value)
-        return dt_util.utc_from_timestamp(timestamp)
-    except (ValueError, TypeError):
-        # Try to parse datetime string to datetime object
-        try:
-            return dt_util.parse_datetime(value, raise_on_error=True)
-        except (ValueError, TypeError):
-            if default is _SENTINEL:
-                # Return None on string input
-                # to ensure backwards compatibility with HA Core 2024.1 and before.
-                if isinstance(value, str):
-                    return None
-                raise_no_default("as_datetime", value)
-            return default
-
-
-def as_timedelta(value: str) -> timedelta | None:
-    """Parse a ISO8601 duration like 'PT10M' to a timedelta."""
-    return dt_util.parse_duration(value)
-
-
 def merge_response(value: ServiceResponse) -> list[Any]:
     """Merge action responses into single list.
 
@@ -1646,16 +1551,6 @@ def merge_response(value: ServiceResponse) -> list[Any]:
     return response_items
 
 
-def strptime(string, fmt, default=_SENTINEL):
-    """Parse a time string to datetime."""
-    try:
-        return datetime.strptime(string, fmt)
-    except (ValueError, AttributeError, TypeError):
-        if default is _SENTINEL:
-            raise_no_default("strptime", string)
-        return default
-
-
 def fail_when_undefined(value):
     """Filter to force a failure when the value is undefined."""
     if isinstance(value, jinja2.Undefined):
@@ -1708,11 +1603,6 @@ def is_number(value):
     if not math.isfinite(fvalue):
         return False
     return True
-
-
-def _is_datetime(value: Any) -> bool:
-    """Return whether a value is a datetime."""
-    return isinstance(value, datetime)
 
 
 def _is_string_like(value: Any) -> bool:
@@ -1817,94 +1707,6 @@ def random_every_time(context, values):
     this is context-dependent to avoid caching the chosen value.
     """
     return random.choice(values)
-
-
-def today_at(hass: HomeAssistant, time_str: str = "") -> datetime:
-    """Record fetching now where the time has been replaced with value."""
-    if (render_info := render_info_cv.get()) is not None:
-        render_info.has_time = True
-
-    today = dt_util.start_of_local_day()
-    if not time_str:
-        return today
-
-    if (time_today := dt_util.parse_time(time_str)) is None:
-        raise ValueError(
-            f"could not convert {type(time_str).__name__} to datetime: '{time_str}'"
-        )
-
-    return datetime.combine(today, time_today, today.tzinfo)
-
-
-def relative_time(hass: HomeAssistant, value: Any) -> Any:
-    """Take a datetime and return its "age" as a string.
-
-    The age can be in second, minute, hour, day, month or year. Only the
-    biggest unit is considered, e.g. if it's 2 days and 3 hours, "2 days" will
-    be returned.
-    If the input datetime is in the future,
-    the input datetime will be returned.
-
-    If the input are not a datetime object the input will be returned unmodified.
-
-    Note: This template function is deprecated in favor of `time_until`, but is still
-    supported so as not to break old templates.
-    """
-
-    if (render_info := render_info_cv.get()) is not None:
-        render_info.has_time = True
-
-    if not isinstance(value, datetime):
-        return value
-    if not value.tzinfo:
-        value = dt_util.as_local(value)
-    if dt_util.now() < value:
-        return value
-    return dt_util.get_age(value)
-
-
-def time_since(hass: HomeAssistant, value: Any | datetime, precision: int = 1) -> Any:
-    """Take a datetime and return its "age" as a string.
-
-    The age can be in seconds, minutes, hours, days, months and year.
-
-    precision is the number of units to return, with the last unit rounded.
-
-    If the value not a datetime object the input will be returned unmodified.
-    """
-    if (render_info := render_info_cv.get()) is not None:
-        render_info.has_time = True
-
-    if not isinstance(value, datetime):
-        return value
-    if not value.tzinfo:
-        value = dt_util.as_local(value)
-    if dt_util.now() < value:
-        return value
-
-    return dt_util.get_age(value, precision)
-
-
-def time_until(hass: HomeAssistant, value: Any | datetime, precision: int = 1) -> Any:
-    """Take a datetime and return the amount of time until that time as a string.
-
-    The time until can be in seconds, minutes, hours, days, months and years.
-
-    precision is the number of units to return, with the last unit rounded.
-
-    If the value not a datetime object the input will be returned unmodified.
-    """
-    if (render_info := render_info_cv.get()) is not None:
-        render_info.has_time = True
-
-    if not isinstance(value, datetime):
-        return value
-    if not value.tzinfo:
-        value = dt_util.as_local(value)
-    if dt_util.now() > value:
-        return value
-
-    return dt_util.get_time_remaining(value, precision)
 
 
 def iif(
@@ -2089,6 +1891,9 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
             "homeassistant.helpers.template.extensions.CollectionExtension"
         )
         self.add_extension("homeassistant.helpers.template.extensions.CryptoExtension")
+        self.add_extension(
+            "homeassistant.helpers.template.extensions.DateTimeExtension"
+        )
         self.add_extension("homeassistant.helpers.template.extensions.DeviceExtension")
         self.add_extension("homeassistant.helpers.template.extensions.FloorExtension")
         self.add_extension("homeassistant.helpers.template.extensions.LabelExtension")
@@ -2097,11 +1902,7 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         self.add_extension("homeassistant.helpers.template.extensions.StringExtension")
 
         self.globals["apply"] = apply
-        self.globals["as_datetime"] = as_datetime
         self.globals["as_function"] = as_function
-        self.globals["as_local"] = dt_util.as_local
-        self.globals["as_timedelta"] = as_timedelta
-        self.globals["as_timestamp"] = forgiving_as_timestamp
         self.globals["bool"] = forgiving_boolean
         self.globals["combine"] = combine
         self.globals["float"] = forgiving_float
@@ -2110,8 +1911,6 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         self.globals["is_number"] = is_number
         self.globals["merge_response"] = merge_response
         self.globals["pack"] = struct_pack
-        self.globals["strptime"] = strptime
-        self.globals["timedelta"] = timedelta
         self.globals["typeof"] = typeof
         self.globals["unpack"] = struct_unpack
         self.globals["version"] = version
@@ -2119,11 +1918,7 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
 
         self.filters["add"] = add
         self.filters["apply"] = apply
-        self.filters["as_datetime"] = as_datetime
         self.filters["as_function"] = as_function
-        self.filters["as_local"] = dt_util.as_local
-        self.filters["as_timedelta"] = as_timedelta
-        self.filters["as_timestamp"] = forgiving_as_timestamp
         self.filters["bool"] = forgiving_boolean
         self.filters["combine"] = combine
         self.filters["contains"] = contains
@@ -2139,9 +1934,6 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         self.filters["pack"] = struct_pack
         self.filters["random"] = random_every_time
         self.filters["round"] = forgiving_round
-        self.filters["timestamp_custom"] = timestamp_custom
-        self.filters["timestamp_local"] = timestamp_local
-        self.filters["timestamp_utc"] = timestamp_utc
         self.filters["to_json"] = to_json
         self.filters["typeof"] = typeof
         self.filters["unpack"] = struct_unpack
@@ -2149,7 +1941,6 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
 
         self.tests["apply"] = apply
         self.tests["contains"] = contains
-        self.tests["datetime"] = _is_datetime
         self.tests["is_number"] = is_number
         self.tests["string_like"] = _is_string_like
 
@@ -2218,15 +2009,9 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
                 "is_hidden_entity",
                 "is_state_attr",
                 "is_state",
-                "now",
-                "relative_time",
                 "state_attr",
                 "state_translated",
                 "states",
-                "time_since",
-                "time_until",
-                "today_at",
-                "utcnow",
             ]
             hass_filters = [
                 "area_id",
@@ -2253,20 +2038,10 @@ class TemplateEnvironment(ImmutableSandboxedEnvironment):
         self.globals["distance"] = hassfunction(distance)
         self.globals["expand"] = hassfunction(expand)
         self.globals["has_value"] = hassfunction(has_value)
-        self.globals["now"] = hassfunction(now)
-        self.globals["relative_time"] = hassfunction(relative_time)
-        self.globals["time_since"] = hassfunction(time_since)
-        self.globals["time_until"] = hassfunction(time_until)
-        self.globals["today_at"] = hassfunction(today_at)
-        self.globals["utcnow"] = hassfunction(utcnow)
 
         self.filters["closest"] = hassfunction(closest_filter)
         self.filters["expand"] = self.globals["expand"]
         self.filters["has_value"] = self.globals["has_value"]
-        self.filters["relative_time"] = self.globals["relative_time"]
-        self.filters["time_since"] = self.globals["time_since"]
-        self.filters["time_until"] = self.globals["time_until"]
-        self.filters["today_at"] = self.globals["today_at"]
 
         self.tests["has_value"] = hassfunction(has_value, pass_eval_context)
