+++ homeassistant/helpers/template/extensions/__init__.py
@@ -4,6 +4,7 @@ from .areas import AreaExtension
 from .base64 import Base64Extension
 from .collection import CollectionExtension
 from .crypto import CryptoExtension
+from .datetime import DateTimeExtension
 from .devices import DeviceExtension
 from .floors import FloorExtension
 from .labels import LabelExtension
@@ -16,6 +17,7 @@ __all__ = [
     "Base64Extension",
     "CollectionExtension",
     "CryptoExtension",
+    "DateTimeExtension",
     "DeviceExtension",
     "FloorExtension",
     "LabelExtension",
